feat: 히스토리 초기작업
