+++ PickaView/Views/MyPageHistories/MyPageHistoriesViewController.swift
@@ -1,0 +1,16 @@
+//
+//  MyPageHistoriesViewController.swift
+//  PickaView
+//
+//  Created by Ko Minhyuk on 6/11/25.
+//
+
+import UIKit
+
+class MyPageHistoriesViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+}
